added test framework for unit testing

+++ server/models/location.py
@@ -1,5 +1,4 @@
 """location embedded model"""
-from decimal import Decimal
 from enum import Enum
 
 from pydantic import BaseModel, conlist
@@ -16,4 +15,4 @@ class Location(BaseModel):  # pylint:disable=too-few-public-methods
     """location embedded model"""
 
     type: LocationType
-    coordinates: conlist(Decimal, min_items=2, max_items=2)  # type: ignore
+    coordinates: conlist(float, min_items=2, max_items=2)  # type: ignore
